Update README.md
Hi could not find a way to contact you, and the other repo is locked so made this as a pull request. Hisuite proxy will not work properly unless date/time is rolled back to 1st sept 2023, it would be useful for others to know when they are trying to upgrade their firmware thank you for you great work.